# apps/api/openapi/ecology.yaml
openapi: 3.1.0
info:
  title: KFM Governed Ecology API
  version: 0.1.0
  description: >
    Public-safe governed API contract for KFM Ecology time-slice outputs,
    EvidenceBundles, and STAC catalog artifacts.

servers:
  - url: /api
    description: Local governed API root

tags:
  - name: Ecology
    description: Governed ecology publication surfaces

paths:
  /ecology/layers/{id}:
    get:
      tags: [Ecology]
      operationId: getEcologyLayer
      summary: Get a governed ecology layer manifest
      description: >
        Returns a public-safe EcologyLayerManifest used to configure MapLibre layers.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ecology layer manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcologyLayerManifest"
        "404":
          description: Layer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "451":
          description: Denied by governance policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /ecology/timeslices/{id}:
    get:
      tags: [Ecology]
      operationId: getEcologyTimeslice
      summary: Get a governed ecology time-slice publication view
      description: >
        Returns a public-safe time-slice response containing tileset metadata,
        promotion state, catalog references, evidence references, and receipt lineage.
      parameters:
        - name: id
          in: path
          required: true
          description: Time-slice or tileset identifier suffix.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Governed ecology time-slice response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EcologyTimesliceResponse"
        "404":
          description: Time-slice not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "451":
          description: Denied by governance policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /ecology/evidence/{bundle_id}:
    get:
      tags: [Ecology]
      operationId: getEcologyEvidenceBundle
      summary: Get a resolved ecology EvidenceBundle
      description: >
        Returns a public-safe resolved EvidenceBundle. Exact restricted geometry,
        RAW/WORK references, and internal-only fields must not be exposed.
      parameters:
        - name: bundle_id
          in: path
          required: true
          description: EvidenceBundle identifier suffix.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Ecology EvidenceBundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceBundle"
        "404":
          description: EvidenceBundle not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "451":
          description: EvidenceBundle denied by governance policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /ecology/catalog/stac:
    get:
      tags: [Ecology]
      operationId: getEcologyStacCatalog
      summary: Get the governed ecology STAC Catalog
      description: >
        Returns the public-safe STAC Catalog generated by the ecology publication pipeline.
      responses:
        "200":
          description: STAC Catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StacCatalog"
        "404":
          description: Catalog not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    EcologyLayerManifest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - object_type
        - layer_id
        - title
        - source_type
        - tiles_url
        - source_layer
        - stac_catalog_ref
        - stac_collection_ref
        - evidence_bundle_ref
        - promotion_decision_ref
        - run_receipt_ref
        - minzoom
        - maxzoom
        - allowed_fields
        - public_safe
      properties:
        schema_version:
          type: string
          const: v1
        object_type:
          type: string
          const: EcologyLayerManifest
        layer_id:
          type: string
        title:
          type: string
        description:
          type: string
        source_type:
          type: string
          enum: [vector, raster, raster-dem, geojson]
        tiles_url:
          type: string
        source_layer:
          type: string
        stac_catalog_ref:
          type: string
        stac_collection_ref:
          type: string
        stac_item_ref:
          type: string
        evidence_bundle_ref:
          type: string
        promotion_decision_ref:
          type: string
        run_receipt_ref:
          type: string
        minzoom:
          type: integer
        maxzoom:
          type: integer
        bounds:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: number
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        allowed_fields:
          type: array
          items:
            type: string
        public_safe:
          type: boolean
          const: true
        sensitivity:
          type: string
        rights_status:
          type: string
        policy_label:
          type: string
        limitations:
          type: array
          items:
            type: string

    EcologyTimesliceResponse:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - object_type
        - timeslice_id
        - tileset_metadata
        - promotion
        - evidence_bundle_ref
        - run_receipt_ref
        - catalog
        - policy
      properties:
        schema_version:
          type: string
          const: v1
        object_type:
          type: string
          const: EcologyTimesliceResponse
        timeslice_id:
          type: string
          pattern: "^kfm://tileset/ecology/.+"
        tileset_metadata:
          $ref: "#/components/schemas/TilesetMetadata"
        promotion:
          $ref: "#/components/schemas/PromotionSummary"
        evidence_bundle_ref:
          type: string
          pattern: "^kfm://evidence/ecology/.+"
        run_receipt_ref:
          type: string
          pattern: "^kfm://receipt/run/ecology/.+"
        catalog:
          $ref: "#/components/schemas/CatalogRefs"
        policy:
          $ref: "#/components/schemas/PolicySummary"

    TilesetMetadata:
      type: object
      additionalProperties: true
      required:
        - scheme
        - bounds
        - minzoom
        - maxzoom
        - time_start
        - time_end
        - expected_tile_count
        - produced_tile_count
        - kfm
        - evidence_bundle_url
        - provisional
        - allowed_fields
      properties:
        scheme:
          type: string
          const: xyz
        bounds:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: number
        minzoom:
          type: integer
          minimum: 0
        maxzoom:
          type: integer
          minimum: 0
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        expected_tile_count:
          type: integer
          minimum: 0
        produced_tile_count:
          type: integer
          minimum: 0
        kfm:
          type: object
          additionalProperties: true
          required: [spec_hash]
          properties:
            spec_hash:
              type: string
        evidence_bundle_url:
          type: string
          pattern: "^kfm://evidence/ecology/.+"
        provisional:
          type: boolean
        allowed_fields:
          type: array
          items:
            type: string

    PromotionSummary:
      type: object
      additionalProperties: false
      required:
        - decision_ref
        - decision
      properties:
        decision_ref:
          type: string
          pattern: "^kfm://promotion/ecology/.+"
        decision:
          type: string
          enum: [PROMOTE, REVIEW, DENY]
        requires_steward:
          type: boolean

    CatalogRefs:
      type: object
      additionalProperties: false
      required:
        - stac_catalog
        - stac_collection
        - stac_item
      properties:
        stac_catalog:
          type: string
        stac_collection:
          type: string
        stac_item:
          type: string

    PolicySummary:
      type: object
      additionalProperties: false
      required:
        - decision
        - policy_ref
      properties:
        decision:
          type: string
          enum: [allow, hold, deny]
        policy_ref:
          type: string
        reasons:
          type: array
          items:
            type: string

    EvidenceBundle:
      type: object
      additionalProperties: true
      required:
        - schema_version
        - object_type
        - bundle_id
        - domain
        - resolved
        - policy_label
        - rights_status
        - sensitivity
        - spec_hash
      properties:
        schema_version:
          type: string
          const: v1
        object_type:
          type: string
          const: EvidenceBundle
        bundle_id:
          type: string
          pattern: "^kfm://evidence/ecology/.+"
        domain:
          type: string
          const: ecology
        resolved:
          type: boolean
        evidence_bundle_resolved:
          type: boolean
        policy_label:
          type: string
        rights_status:
          type: string
          enum: [public, open, controlled, restricted, unknown]
        sensitivity:
          type: string
          enum: [public, generalize, restricted, review_required]
        spec_hash:
          type: string
          pattern: "^sha256:[a-fA-F0-9]{64}$"

    StacCatalog:
      type: object
      additionalProperties: true
      required:
        - stac_version
        - type
        - id
        - title
        - description
        - links
        - kfm:domain
        - kfm:collection_count
      properties:
        stac_version:
          type: string
          const: 1.0.0
        type:
          type: string
          const: Catalog
        id:
          type: string
        title:
          type: string
        description:
          type: string
        links:
          type: array
          items:
            type: object
            additionalProperties: true
        kfm:domain:
          type: string
          const: ecology
        kfm:collection_count:
          type: integer
          minimum: 1

    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - object_type
        - outcome
        - message
      properties:
        object_type:
          type: string
          const: ErrorEnvelope
        outcome:
          type: string
          enum: [ERROR, DENY, ABSTAIN]
        message:
          type: string
        reasons:
          type: array
          items:
            type: string
        receipt_ref:
          type: string